requirements.txt
pytest
requests
allure-pytest
toml
